Tests/LibWeb/Text/expected/hit_testing/hit-test-preserves-pending-repaint.txt

needsRepaint after mouse move: true
